refactor: revamp folder structure
