+++ frontend/src/assets/icons/index.ts
@@ -1,0 +1,21 @@
+import BeatblockLogo from "./BEATBLOCK.svg";
+import BeatblockLogoSmall from "./beatblock-icon-only.svg";
+import MenuIcon from "./menu.svg";
+import UploadIcon from "./upload.svg";
+import DefaultAudioImage from "./default-audio-image.svg";
+import TrebleClefIcon from "./treble-clef.svg";
+import DrumIcon from "./drums.svg";
+import MusicIcon from "./music.svg";
+import SliderIcon from "./sliders.svg";
+
+export {
+  BeatblockLogo,
+  BeatblockLogoSmall,
+  MenuIcon,
+  UploadIcon,
+  DefaultAudioImage,
+  TrebleClefIcon,
+  DrumIcon,
+  MusicIcon,
+  SliderIcon,
+};
