+++ frontend/src/assets/textures/index.ts
@@ -1,0 +1,5 @@
+import ArrowTexture from './arrow-texture.svg';
+import GridTexture from './grid-texture.svg';
+import LineTexture from './line-texture.svg';
+
+export { ArrowTexture, GridTexture, LineTexture };
